Done with the main part, need to run official tests and shorten

+++ 5_Push_Swap/push_swap_main.c
@@ -63,7 +63,7 @@ int	main(int argc, char **argv)
 	i = 0;
 	while (i < array_size)
 	{
-		if (((numbers_array[i]) < INT_MIN) || ((numbers_array[i]) > INT_MAX))
+		if (((numbers_array[i]) < (long int)INT_MIN) || ((numbers_array[i]) > (long int)INT_MAX))
 			clean_early_exit(ERROR_MESSSAGE_03, numbers_array, true);
 		i++;
 	}
